requirements.txt
torch
torchvision
streamlit
plotly
numpy
Pillow
scikit-image
scikit-learn
